refactor(node)!: Move `Yarn` into its own dedicated directory
Align with `pnpm` and `yarn2`.
