+++ plugins/package-managers/node/src/main/kotlin/yarn/YarnDependencyHandler.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.plugins.packagemanagers.node.utils
+package org.ossreviewtoolkit.plugins.packagemanagers.node.yarn
 
 import java.io.File
 
@@ -27,7 +27,7 @@ import org.ossreviewtoolkit.model.Package
 import org.ossreviewtoolkit.model.PackageLinkage
 import org.ossreviewtoolkit.model.Project
 import org.ossreviewtoolkit.model.utils.DependencyHandler
-import org.ossreviewtoolkit.plugins.packagemanagers.node.Yarn
+import org.ossreviewtoolkit.plugins.packagemanagers.node.utils.parsePackage
 
 /**
  * A data class storing information about a specific NPM module and its dependencies.
