+++ plugins/package-managers/node/src/test/kotlin/utils/NpmDependencyHandlerTest.kt
@@ -32,6 +32,8 @@ import org.ossreviewtoolkit.model.PackageLinkage
 import org.ossreviewtoolkit.model.config.AnalyzerConfiguration
 import org.ossreviewtoolkit.model.config.RepositoryConfiguration
 import org.ossreviewtoolkit.plugins.packagemanagers.node.Npm
+import org.ossreviewtoolkit.plugins.packagemanagers.node.yarn.NpmModuleInfo
+import org.ossreviewtoolkit.plugins.packagemanagers.node.yarn.YarnDependencyHandler
 import org.ossreviewtoolkit.utils.test.USER_DIR
 
 class NpmDependencyHandlerTest : StringSpec({
