+++ plugins/package-managers/node/src/test/kotlin/yarn/YarnTest.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.plugins.packagemanagers.node
+package org.ossreviewtoolkit.plugins.packagemanagers.node.yarn
 
 import io.kotest.core.spec.style.WordSpec
 import io.kotest.matchers.nulls.beNull
